Fixed bug that caused error when cancelling order with no items

+++ Pages/OrderManagement/CreateOrder.cshtml.cs
@@ -68,8 +68,11 @@ namespace FutureFridges.Pages.OrderManagement
 
         public IActionResult OnGetCancelOrder (Guid uid)
         {
-            __OrderController.DeleteOrder(uid);
-
+            if (uid != Guid.Empty)
+            {
+                __OrderController.DeleteOrder(uid);
+            }
+            
             return RedirectToPage("../Index");
         }
